+++ services/preview/signalr/mgmt/2020-07-01-preview/signalr/CHANGELOG.md
@@ -1,2 +1,7 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. ResourceSku.MarshalJSON() ([]byte, error)