+++ services/signalr/mgmt/2020-05-01/signalr/CHANGELOG.md
@@ -1,2 +1,12 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. RequestType.Trace
+
+### New Funcs
+
+1. CreateOrUpdateProperties.MarshalJSON() ([]byte, error)
+1. ResourceSku.MarshalJSON() ([]byte, error)